requirements.txt
# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Telegram Bot
python-telegram-bot==21.0

# Google Gemini (using new SDK)
google-genai==0.2.2

# Excel Export
openpyxl==3.1.2

# Image Processing
Pillow==10.2.0

# HTTP Client
httpx==0.27.0
aiohttp==3.9.1

# Environment Variables
python-dotenv==1.0.0

# Email
aiosmtplib==3.0.1

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
python-dateutil==2.8.2
